fix: fix race condition in adding APIs
When a new module adds existing APIs, it shoud register a listener for new APIs first before it asks
for the existing list.  There was a small window between when the API list was queried and acted on
and when the listener was registered.  This window allowed a new API to be registered but not seen
by this process.

closes BSUM-58