fix(module_load): make dependency errors useful
The dependency error will now display the dependent module and the version
criteria it must meet.

There is also a change to the "module info" pop-up in that it shows module
version as well.  Ideally, it should also show dependency information but that
will be for a future dev to fix.

closes BITS-244